+++ main.go
@@ -1,125 +1,17 @@
 package main
 
 import (
-	"fmt"
+	"ginLearn/common"
 	"github.com/gin-gonic/gin"
 	_ "github.com/go-sql-driver/mysql"
-	"github.com/jinzhu/gorm"
-	"log"
-	"math/rand"
-	"net/http"
-	"time"
 )
 
-type User struct {
-	gorm.Model
-	Name      string `gorm:"type:varchar(20);not null"`
-	Telephone string `gorm:"type:varchar(20);not null;unique"`
-	Password  string `gorm:"size:255;not null"`
-}
-
 func main() {
-	db := InitDB()
+	db := common.InitDB()
 	defer db.Close()
 
 	r := gin.Default()
-	r.POST("/api/auth/register", func(c *gin.Context) {
-		name := c.PostForm("name")
-		telephone := c.PostForm("telephone")
-		password := c.PostForm("password")
+	r = CollectRoute(r)
 
-		if len(telephone) != 11 {
-			c.JSON(http.StatusUnprocessableEntity, map[string]interface{}{ //gin.H
-				"code": 442,
-				"msg":  "手机号错误,手机号必须为11位!",
-			})
-			return
-		}
-
-		if len(password) < 6 {
-			c.JSON(http.StatusUnprocessableEntity, gin.H{
-				"code": 442,
-				"msg":  "密码至少为6位!",
-			})
-			return
-		}
-
-		if len(name) == 0 {
-			name = RandomString(10)
-		}
-
-		if isTelephoneExist(db, telephone) {
-			c.JSON(http.StatusUnprocessableEntity, gin.H{
-				"code": 442,
-				"msg":  "用户已存在!",
-			})
-			return
-		}
-
-		newUser := User{
-			Name:      name,
-			Telephone: telephone,
-			Password:  password,
-		}
-		db.Create(&newUser)
-
-		log.Println(name, telephone, password)
-
-		c.JSON(200, gin.H{
-			"code": 200,
-			"msg":  "注册成功!",
-		})
-	})
 	_ = r.Run() // listen and serve on 0.0.0.0:8080 (for windows "localhost:8080")
-}
-
-func isTelephoneExist(db *gorm.DB, telephone string) bool {
-	var user User
-	db.Where("telephone = ?", telephone).First(&user)
-
-	if user.ID != 0 {
-		return true
-	}
-
-	return false
-}
-
-func RandomString(n int) string {
-	var letters = []byte("abcdefghtyuikllwetyibvhjigkgggkgju")
-	result := make([]byte, n)
-	// 设置种子，不然每次都会随机成0
-	rand.Seed(time.Now().Unix())
-
-	for i := range result {
-		// 生成 letters的长度(n) 到 (n-1) 的随机数（输出 int类型）
-		result[i] = letters[rand.Intn(len(letters))]
-	}
-
-	return string(result)
-}
-
-func InitDB() *gorm.DB {
-	driverName := "mysql"
-	host := "localhost"
-	port := "3306"
-	database := "ginessential"
-	username := "root"
-	password := "root"
-	charset := "utf8"
-	args := fmt.Sprintf("%s:%s@tcp(%s:%s)/%s?charset=%s&parseTime=true",
-		username,
-		password,
-		host,
-		port,
-		database,
-		charset,
-	)
-	db, err := gorm.Open(driverName, args)
-	if err != nil {
-		panic("连接数据库失败" + err.Error())
-	}
-	log.Println(db.HasTable("users"))
-	db.AutoMigrate(&User{})
-
-	return db
 }
