+++ model/user.go
@@ -1,0 +1,10 @@
+package model
+
+import "github.com/jinzhu/gorm"
+
+type User struct {
+	gorm.Model
+	Name      string `gorm:"type:varchar(20);not null"`
+	Telephone string `gorm:"type:varchar(20);not null;unique"`
+	Password  string `gorm:"size:255;not null"`
+}
