+++ routes.go
@@ -1,0 +1,12 @@
+package main
+
+import (
+	"ginLearn/controller"
+	"github.com/gin-gonic/gin"
+)
+
+func CollectRoute(r *gin.Engine) *gin.Engine {
+	r.POST("api/auth/register", controller.Register)
+
+	return r
+}
